+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Changed
 * Moved `BBox` to `pointy` crate
-* Updated `protobuf` dependency to version 2.25
 * `GeomEncoder::point` / `add_point` are now fallible (float to int errors)
+* `GeomEncoder` now has a `Float` type parameter (`f32` or `f64`)
+* Updated `protobuf` dependency to version 2.25
 
 ## [0.7.0] - 2020-09-29
 ### Changed